+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Loading screen shown while the app is starting
+- Exercises page: Filtering by muscles differentiating between primary and secondary muscles
 - Catalog exercise page: Order of the properties of an exercise, showing the muscles first
 
 ### Fixed